Make read queries portable across Postgres and SQLite
Tests run against in-memory SQLite while the app runs on Postgres, so the
Postgres-only constructs are replaced: ILIKE -> LOWER(...) LIKE LOWER(...),
STRING_AGG species_list built in Python (_species_list_map), calendar_month
EXTRACT(...::date) parsed from the date string, and ::numeric casts dropped.

list_sessions now orders NULL start_times last portably
(ORDER BY date DESC, (start_time IS NULL), start_time DESC).

+++ fishing_log/analytics.py
@@ -258,9 +258,9 @@ def by_species() -> pd.DataFrame:
             f.species,
             COUNT(*) AS total_caught,
             COUNT(DISTINCT f.session_id) AS sessions_present,
-            ROUND(AVG(NULLIF(f.length, 0))::numeric, 2) AS avg_length,
+            ROUND(AVG(NULLIF(f.length, 0)), 2) AS avg_length,
             MAX(f.length) AS max_length,
-            ROUND(AVG(NULLIF(f.weight, 0))::numeric, 2) AS avg_weight,
+            ROUND(AVG(NULLIF(f.weight, 0)), 2) AS avg_weight,
             MAX(f.weight) AS max_weight
         FROM fish f JOIN sessions s ON s.id = f.session_id
         WHERE s.user_email = :email
